{"schema":"decodex/commit/1","summary":"Move scroll stitching into capture core","authority":"manual"}
