+++ Cargo.toml
@@ -31,7 +31,6 @@ objc2-core-media         = { version = "0.3" }
 objc2-core-video         = { version = "0.3" }
 objc2-foundation         = { version = "0.3" }
 objc2-screen-capture-kit = { version = "0.3" }
-objc2-vision             = { version = "0.3" }
 png                      = { version = "0.18" }
 serde                    = { version = "1.0", features = ["derive"] }
 thiserror                = { version = "2.0" }
