+++ packages/rsnap-capture-core/src/scroll_capture/bench_support.rs
