+++ packages/rsnap-capture-core/src/scroll_capture/downward_committed_recovery.rs
