+++ packages/rsnap-capture-core/src/scroll_capture/downward_recovery.rs
