+++ packages/rsnap-capture-core/src/scroll_capture/downward_registration.rs
