+++ packages/rsnap-capture-core/src/scroll_capture/downward_resolution.rs
