+++ packages/rsnap-capture-core/src/scroll_capture/fingerprint.rs
