+++ packages/rsnap-capture-core/src/scroll_capture/image_stack.rs
