+++ packages/rsnap-capture-core/src/scroll_capture/informative_span.rs
