+++ packages/rsnap-capture-core/src/scroll_capture/logging.rs
