+++ packages/rsnap-capture-core/src/scroll_capture/resume_frontier.rs
