+++ packages/rsnap-capture-core/src/scroll_capture/sample_input.rs
