+++ packages/rsnap-capture-core/src/scroll_capture/sample_registration.rs
