+++ packages/rsnap-capture-core/src/scroll_capture/state_access.rs
