+++ packages/rsnap-capture-core/src/scroll_capture/support.rs
