+++ packages/rsnap-capture-core/src/scroll_capture/test_support.rs
