+++ packages/rsnap-capture-core/src/scroll_capture/upward_input.rs
